refactor(ir): extract expand_mixed_kernel utilities into reusable modules Split the 2563-line monolithic expand_mixed_kernel_pass.cpp into focused modules, reducing it to 882 lines while preserving all functionality: - core_affinity.h/cpp: CoreAffinity enum and op classification - cross_core_pipe.h/cpp: pipe metadata collection and auto setup - dead_code_elimination.h/cpp: statement-level DCE - loop_state_repair.h/cpp: iter_arg stripping, init value fixup,   dangling yield repair, and FinalizeSplitCoreBody - tpop_chain_normalizer.h/cpp: tpop/tfree chain ordering and tfree   insertion - mixed_kernel_expanded_verifier.cpp: property verifier for post-pass   validation Also fix duplicate loop_state_repair.cpp entry in CMakeLists.txt.
