+++ CMakeLists.txt
@@ -158,7 +158,12 @@ set(PYPTO_SOURCES
     src/ir/transforms/unroll_loops_pass.cpp
     src/ir/transforms/structural_equal.cpp
     src/ir/transforms/structural_hash.cpp
+    src/ir/transforms/utils/core_affinity.cpp
+    src/ir/transforms/utils/cross_core_pipe.cpp
     src/ir/transforms/utils/deep_clone_utils.cpp
+    src/ir/transforms/utils/dead_code_elimination.cpp
+    src/ir/transforms/utils/loop_state_repair.cpp
+    src/ir/transforms/utils/tpop_chain_normalizer.cpp
     src/ir/transforms/utils/normalize_stmt_structure.cpp
     src/ir/transforms/utils/parent_stmt_analysis.cpp
     src/ir/transforms/utils/transform_utils.cpp
@@ -172,6 +177,7 @@ set(PYPTO_SOURCES
     # IR - Verifier
     src/ir/verifier/verifier.cpp
     src/ir/verifier/property_verifier_registry.cpp
+    src/ir/verifier/mixed_kernel_expanded_verifier.cpp
     src/ir/verifier/verify_ssa_pass.cpp
     src/ir/verifier/verify_no_nested_call_pass.cpp
     src/ir/verifier/verify_break_continue_pass.cpp
