+++ src/ir/transforms/expand_mixed_kernel_pass.cpp
@@ -37,476 +37,44 @@
 #include "pypto/ir/scalar_expr.h"
 #include "pypto/ir/span.h"
 #include "pypto/ir/stmt.h"
-#include "pypto/ir/transforms/base/visitor.h"
 #include "pypto/ir/transforms/pass_properties.h"
 #include "pypto/ir/transforms/passes.h"
+#include "pypto/ir/transforms/utils/core_affinity.h"
+#include "pypto/ir/transforms/utils/cross_core_pipe.h"
+#include "pypto/ir/transforms/utils/dead_code_elimination.h"
 #include "pypto/ir/transforms/utils/deep_clone_utils.h"
+#include "pypto/ir/transforms/utils/loop_state_repair.h"
 #include "pypto/ir/transforms/utils/scope_outline_utils.h"
+#include "pypto/ir/transforms/utils/tpop_chain_normalizer.h"
 #include "pypto/ir/transforms/utils/transform_utils.h"
 #include "pypto/ir/type.h"
-#include "pypto/ir/verifier/verifier.h"
 
 namespace pypto {
 namespace ir {
 
 namespace {
 
-// ============================================================================
-// Core Affinity Classification
-// ============================================================================
-
-enum class CoreAffinity { CUBE, VECTOR, SHARED, MIXED, BOUNDARY };
-
-enum class PipeDirection { C2V = 1, V2C = 2 };
-enum class CoreSide { AIC, AIV };
-
-constexpr int kDirMaskC2V = static_cast<int>(PipeDirection::C2V);
-constexpr int kDirMaskV2C = static_cast<int>(PipeDirection::V2C);
-constexpr int kAutoBufferBase = -1;
-
-CoreAffinity CombineAffinity(CoreAffinity a, CoreAffinity b) {
-  if (a == b) return a;
-  if (a == CoreAffinity::SHARED) return b;
-  if (b == CoreAffinity::SHARED) return a;
-  return CoreAffinity::MIXED;
-}
-
-bool IsCubeOp(const std::string& name) {
-  static const std::unordered_set<std::string> cube_ops = {
-      "tile.matmul",   "tile.matmul_acc", "tile.matmul_bias", "tile.gemv",
-      "tile.gemv_acc", "tile.gemv_bias",  "tile.batch_matmul"};
-  return cube_ops.count(name) > 0;
-}
-
-bool IsCubeMemorySpace(MemorySpace ms) { return ms != MemorySpace::DDR && ms != MemorySpace::Vec; }
-
-/// Get target_memory from the first tile-typed argument of a Call.
-/// Returns nullopt if no tile-typed Var argument is found.
-std::optional<MemorySpace> GetFirstTileArgMemory(const CallPtr& call) {
-  for (const auto& arg : call->args_) {
-    if (auto var = std::dynamic_pointer_cast<const Var>(arg)) {
-      if (auto tile_type = std::dynamic_pointer_cast<const TileType>(var->GetType())) {
-        return tile_type->memory_space_;
-      }
-    }
-  }
-  return std::nullopt;
-}
+using core_affinity::ClassifyCallAffinity;
+using core_affinity::ClassifyMoveDirection;
+using core_affinity::CombineAffinity;
+using core_affinity::CoreAffinity;
+using core_affinity::CoreSide;
+using core_affinity::CVBoundaryMove;
+using core_affinity::CVDirection;
+using core_affinity::IsCubeMemorySpace;
+using cross_core_pipe::BuildAutomaticPipeSetup;
+using cross_core_pipe::PrependPipeSetup;
+using loop_repair::BuildDefMap;
+using loop_repair::FinalizeSplitCoreBody;
+using loop_repair::MakeBody;
+using tpop_chain::NormalizeTpopChains;
 
 // ============================================================================
-// CV Boundary Move Detection
-// ============================================================================
-
-enum class CVDirection { NONE, CUBE_TO_VECTOR, VECTOR_TO_CUBE };
-
-/// Classify whether a Call is a CV-boundary tile.move.
-/// Returns CUBE_TO_VECTOR if source is cube memory and target is vector memory.
-/// Returns VECTOR_TO_CUBE if source is vector memory and target is cube memory.
-/// Returns NONE for non-tile.move calls or same-side moves.
-CVDirection ClassifyMoveDirection(const CallPtr& call) {
-  if (!call || !call->op_) return CVDirection::NONE;
-
-  auto op = std::dynamic_pointer_cast<const Op>(call->op_);
-  if (!op || op->name_ != "tile.move") return CVDirection::NONE;
-
-  auto src_memory = GetFirstTileArgMemory(call);
-  if (!src_memory.has_value()) return CVDirection::NONE;
-
-  // target_memory kwarg is always present on tile.move (ensured by InferTileTargetMemory)
-  std::optional<MemorySpace> target_memory;
-  for (const auto& [key, value] : call->kwargs_) {
-    if (key == "target_memory") {
-      target_memory = AnyCast<MemorySpace>(value, "target_memory");
-      break;
-    }
-  }
-  INTERNAL_CHECK(target_memory.has_value()) << "Internal error: tile.move missing target_memory kwarg";
-
-  bool src_cube = IsCubeMemorySpace(src_memory.value());
-  bool tgt_cube = IsCubeMemorySpace(target_memory.value());
-  if (src_cube && !tgt_cube) return CVDirection::CUBE_TO_VECTOR;
-  if (!src_cube && tgt_cube) return CVDirection::VECTOR_TO_CUBE;
-  return CVDirection::NONE;
-}
-
-// ============================================================================
-// Core Affinity Classification (call-level)
-// ============================================================================
-
-CoreAffinity ClassifyCallAffinity(const CallPtr& call) {
-  if (!call || !call->op_) return CoreAffinity::SHARED;
-
-  // GlobalVar call (function call) is SHARED
-  if (std::dynamic_pointer_cast<const GlobalVar>(call->op_)) {
-    return CoreAffinity::SHARED;
-  }
-
-  auto op = std::dynamic_pointer_cast<const Op>(call->op_);
-  if (!op) return CoreAffinity::SHARED;
-
-  const auto& name = op->name_;
-
-  // Cube ops (matmul, gemv, etc.)
-  if (IsCubeOp(name)) return CoreAffinity::CUBE;
-
-  // tile.move: CV boundary moves get BOUNDARY affinity;
-  // non-boundary moves are classified by source tile memory space.
-  if (name == "tile.move") {
-    auto dir = ClassifyMoveDirection(call);
-    if (dir != CVDirection::NONE) return CoreAffinity::BOUNDARY;
-    auto ms = GetFirstTileArgMemory(call);
-    if (ms.has_value() && IsCubeMemorySpace(ms.value())) return CoreAffinity::CUBE;
-    return CoreAffinity::VECTOR;
-  }
-
-  // tile.store, tile.reshape: classified by source tile memory space.
-  static const std::unordered_set<std::string> tile_arg_classified_ops = {"tile.store", "tile.reshape"};
-  if (tile_arg_classified_ops.count(name)) {
-    auto ms = GetFirstTileArgMemory(call);
-    if (ms.has_value() && IsCubeMemorySpace(ms.value())) return CoreAffinity::CUBE;
-    return CoreAffinity::VECTOR;
-  }
-
-  // tile.load: classify by target_memory kwarg (no tile input to inspect)
-  if (name == "tile.load") {
-    for (const auto& [key, value] : call->kwargs_) {
-      if (key == "target_memory") {
-        return IsCubeMemorySpace(AnyCast<MemorySpace>(value, "target_memory")) ? CoreAffinity::CUBE
-                                                                               : CoreAffinity::VECTOR;
-      }
-    }
-    return CoreAffinity::VECTOR;  // default target_memory is Vec
-  }
-
-  // Cross-core tile ops: must be checked before the generic tile.* branch.
-  // AIV-side ops are VECTOR, AIC-side ops are CUBE.
-  static const std::unordered_set<std::string> vector_cross_core_ops = {
-      "system.aiv_initialize_pipe", "system.tpush_to_aic", "system.tfree_to_aic", "tile.tpush_to_aic",
-      "tile.tpop_from_aic"};
-  if (vector_cross_core_ops.count(name)) return CoreAffinity::VECTOR;
-
-  static const std::unordered_set<std::string> cube_cross_core_ops = {
-      "system.aic_initialize_pipe", "system.tfree_to_aiv", "system.tpush_to_aiv", "tile.tpush_to_aiv",
-      "tile.tpop_from_aiv"};
-  if (cube_cross_core_ops.count(name)) return CoreAffinity::CUBE;
-
-  // Other tile.* ops are vector
-  if (name.substr(0, 5) == "tile.") return CoreAffinity::VECTOR;
-
-  return CoreAffinity::SHARED;
-}
-
-// ============================================================================
-// Flatten body / make body helpers
+// Flatten body helper
 // ============================================================================
 
 // Use the shared utility; local alias preserves call sites.
 const auto& FlattenBody = transform_utils::FlattenToStmts;
-
-StmtPtr MakeBody(const std::vector<StmtPtr>& stmts, const Span& span) {
-  return SeqStmts::Flatten(std::vector<StmtPtr>(stmts), span);
-}
-
-struct PipeDirectionMetadata {
-  bool has_ops = false;
-  bool has_inconsistent_slot_size = false;
-  std::optional<int64_t> slot_size_bytes;
-  std::vector<int64_t> observed_slot_sizes;
-};
-
-struct CrossCorePipeMetadata {
-  PipeDirectionMetadata c2v;
-  PipeDirectionMetadata v2c;
-  bool has_reserve_buffer = false;
-  bool has_import_peer_buffer = false;
-  bool has_aic_initialize_pipe = false;
-  bool has_aiv_initialize_pipe = false;
-
-  [[nodiscard]] bool HasCrossCoreOps() const { return c2v.has_ops || v2c.has_ops; }
-  [[nodiscard]] bool HasAnySetup() const {
-    return has_reserve_buffer || has_import_peer_buffer || has_aic_initialize_pipe || has_aiv_initialize_pipe;
-  }
-};
-
-std::optional<int64_t> TryGetConstIntValue(const ExprPtr& expr) {
-  auto const_int = std::dynamic_pointer_cast<const ConstInt>(expr);
-  if (!const_int || const_int->value_ < 0) return std::nullopt;
-  return const_int->value_;
-}
-
-std::optional<int64_t> TryGetTileSlotSizeBytes(const TypePtr& type) {
-  auto tile_type = std::dynamic_pointer_cast<const TileType>(type);
-  if (!tile_type) return std::nullopt;
-
-  int64_t element_count = 1;
-  for (const auto& dim : tile_type->shape_) {
-    auto dim_value = TryGetConstIntValue(dim);
-    if (!dim_value.has_value()) return std::nullopt;
-    CHECK(*dim_value == 0 || element_count <= std::numeric_limits<int64_t>::max() / *dim_value)
-        << "Tile element count overflow while inferring cross-core slot size";
-    element_count *= *dim_value;
-  }
-
-  const int64_t bit_width = static_cast<int64_t>(tile_type->dtype_.GetBit());
-  CHECK(bit_width > 0) << "Unsupported dtype for cross-core slot size inference: "
-                       << tile_type->dtype_.ToString();
-  CHECK(element_count <= (std::numeric_limits<int64_t>::max() - 7) / bit_width)
-      << "Tile byte size overflow while inferring cross-core slot size";
-  return (element_count * bit_width + 7) / 8;
-}
-
-void RecordObservedSlotSize(PipeDirectionMetadata& metadata, int64_t slot_size) {
-  metadata.has_ops = true;
-  if (std::find(metadata.observed_slot_sizes.begin(), metadata.observed_slot_sizes.end(), slot_size) ==
-      metadata.observed_slot_sizes.end()) {
-    metadata.observed_slot_sizes.push_back(slot_size);
-  }
-  if (!metadata.slot_size_bytes.has_value()) {
-    metadata.slot_size_bytes = slot_size;
-    return;
-  }
-  if (metadata.slot_size_bytes.value() != slot_size) {
-    metadata.has_inconsistent_slot_size = true;
-  }
-}
-
-void RecordTileSlotSize(PipeDirectionMetadata& metadata, const TypePtr& type) {
-  metadata.has_ops = true;
-  auto slot_size = TryGetTileSlotSizeBytes(type);
-  if (slot_size.has_value()) {
-    RecordObservedSlotSize(metadata, slot_size.value());
-  }
-}
-
-void MergeDirectionMetadata(PipeDirectionMetadata& dst, const PipeDirectionMetadata& src) {
-  dst.has_ops = dst.has_ops || src.has_ops;
-  dst.has_inconsistent_slot_size = dst.has_inconsistent_slot_size || src.has_inconsistent_slot_size;
-  for (int64_t slot_size : src.observed_slot_sizes) {
-    RecordObservedSlotSize(dst, slot_size);
-  }
-}
-
-CrossCorePipeMetadata MergeCrossCorePipeMetadata(const CrossCorePipeMetadata& lhs,
-                                                 const CrossCorePipeMetadata& rhs) {
-  CrossCorePipeMetadata merged;
-  MergeDirectionMetadata(merged.c2v, lhs.c2v);
-  MergeDirectionMetadata(merged.c2v, rhs.c2v);
-  MergeDirectionMetadata(merged.v2c, lhs.v2c);
-  MergeDirectionMetadata(merged.v2c, rhs.v2c);
-  merged.has_reserve_buffer = lhs.has_reserve_buffer || rhs.has_reserve_buffer;
-  merged.has_import_peer_buffer = lhs.has_import_peer_buffer || rhs.has_import_peer_buffer;
-  merged.has_aic_initialize_pipe = lhs.has_aic_initialize_pipe || rhs.has_aic_initialize_pipe;
-  merged.has_aiv_initialize_pipe = lhs.has_aiv_initialize_pipe || rhs.has_aiv_initialize_pipe;
-  return merged;
-}
-
-int BuildDirMask(const CrossCorePipeMetadata& metadata) {
-  int dir_mask = 0;
-  if (metadata.c2v.has_ops) dir_mask |= kDirMaskC2V;
-  if (metadata.v2c.has_ops) dir_mask |= kDirMaskV2C;
-  return dir_mask;
-}
-
-int GetSlotNumForDirMask(int dir_mask) { return dir_mask == (kDirMaskC2V | kDirMaskV2C) ? 4 : 8; }
-
-std::optional<int64_t> GetCommonSlotSizeBytes(const CrossCorePipeMetadata& metadata) {
-  std::optional<int64_t> common_slot_size;
-  for (const auto* direction : {&metadata.c2v, &metadata.v2c}) {
-    if (!direction->has_ops) continue;
-    if (direction->has_inconsistent_slot_size || !direction->slot_size_bytes.has_value()) {
-      return std::nullopt;
-    }
-    if (!common_slot_size.has_value()) {
-      common_slot_size = direction->slot_size_bytes;
-      continue;
-    }
-    if (common_slot_size.value() != direction->slot_size_bytes.value()) {
-      return std::nullopt;
-    }
-  }
-  return common_slot_size;
-}
-
-std::string BuildPipeBufferName(const std::string& func_name, PipeDirection direction) {
-  return func_name + ((direction == PipeDirection::C2V) ? "_c2v_slot_buffer" : "_v2c_slot_buffer");
-}
-
-CallPtr CreateSystemOpCall(const std::string& op_name,
-                           const std::vector<std::pair<std::string, std::any>>& kwargs, const Span& span) {
-  return OpRegistry::GetInstance().Create(op_name, {}, kwargs, span);
-}
-
-CallPtr CreateReserveBuffer(const std::string& buffer_name, int64_t size_bytes, const Span& span) {
-  CHECK(size_bytes >= 0 && size_bytes <= std::numeric_limits<int>::max())
-      << "Cross-core reserve_buffer size out of range: " << size_bytes;
-  return CreateSystemOpCall("system.reserve_buffer",
-                            {{"name", std::any(buffer_name)},
-                             {"size", std::any(static_cast<int>(size_bytes))},
-                             {"base", std::any(kAutoBufferBase)}},
-                            span);
-}
-
-CallPtr CreateImportPeerBuffer(const std::string& buffer_name, const std::string& peer_func,
-                               const Span& span) {
-  return CreateSystemOpCall("system.import_peer_buffer",
-                            {{"name", std::any(buffer_name)}, {"peer_func", std::any(peer_func)}}, span);
-}
-
-CallPtr CreateInitializePipe(CoreSide side, int dir_mask, int slot_size_bytes, const Span& span) {
-  CHECK(slot_size_bytes >= 0 && slot_size_bytes <= std::numeric_limits<int>::max())
-      << "Cross-core slot_size out of range: " << slot_size_bytes;
-  const std::string op_name =
-      (side == CoreSide::AIC) ? "system.aic_initialize_pipe" : "system.aiv_initialize_pipe";
-  return CreateSystemOpCall(
-      op_name, {{"dir_mask", std::any(dir_mask)}, {"slot_size", std::any(slot_size_bytes)}}, span);
-}
-
-void CollectCrossCorePipeMetadata(const std::vector<StmtPtr>& stmts, CrossCorePipeMetadata& metadata) {
-  for (const auto& stmt : stmts) {
-    auto assign = std::dynamic_pointer_cast<const AssignStmt>(stmt);
-    auto eval = std::dynamic_pointer_cast<const EvalStmt>(stmt);
-    CallPtr call;
-    if (assign) {
-      call = std::dynamic_pointer_cast<const Call>(assign->value_);
-    } else if (eval) {
-      call = std::dynamic_pointer_cast<const Call>(eval->expr_);
-    }
-    auto op = call ? std::dynamic_pointer_cast<const Op>(call->op_) : nullptr;
-    if (op) {
-      const std::string& op_name = op->name_;
-      if (op_name == "system.reserve_buffer") {
-        metadata.has_reserve_buffer = true;
-      } else if (op_name == "system.import_peer_buffer") {
-        metadata.has_import_peer_buffer = true;
-      } else if (op_name == "system.aic_initialize_pipe") {
-        metadata.has_aic_initialize_pipe = true;
-      } else if (op_name == "system.aiv_initialize_pipe") {
-        metadata.has_aiv_initialize_pipe = true;
-      } else if (op_name == "tile.tpush_to_aiv" && call->args_.size() == 1) {
-        RecordTileSlotSize(metadata.c2v, call->args_[0]->GetType());
-      } else if (op_name == "tile.tpush_to_aic" && call->args_.size() == 1) {
-        RecordTileSlotSize(metadata.v2c, call->args_[0]->GetType());
-      } else if (op_name == "tile.tpop_from_aiv" && assign) {
-        RecordTileSlotSize(metadata.v2c, assign->var_->GetType());
-      } else if (op_name == "tile.tpop_from_aic" && assign) {
-        RecordTileSlotSize(metadata.c2v, assign->var_->GetType());
-      }
-    }
-
-    if (auto for_stmt = std::dynamic_pointer_cast<const ForStmt>(stmt)) {
-      CollectCrossCorePipeMetadata(FlattenBody(for_stmt->body_), metadata);
-    } else if (auto if_stmt = std::dynamic_pointer_cast<const IfStmt>(stmt)) {
-      CollectCrossCorePipeMetadata(FlattenBody(if_stmt->then_body_), metadata);
-      if (if_stmt->else_body_.has_value()) {
-        CollectCrossCorePipeMetadata(FlattenBody(if_stmt->else_body_.value()), metadata);
-      }
-    } else if (auto while_stmt = std::dynamic_pointer_cast<const WhileStmt>(stmt)) {
-      CollectCrossCorePipeMetadata(FlattenBody(while_stmt->body_), metadata);
-    }
-  }
-}
-
-CrossCorePipeMetadata CollectDominatingPipeSetupMetadata(const std::vector<StmtPtr>& stmts) {
-  CrossCorePipeMetadata metadata;
-  for (const auto& stmt : stmts) {
-    auto assign = std::dynamic_pointer_cast<const AssignStmt>(stmt);
-    auto eval = std::dynamic_pointer_cast<const EvalStmt>(stmt);
-    CallPtr call;
-    if (assign) {
-      call = std::dynamic_pointer_cast<const Call>(assign->value_);
-    } else if (eval) {
-      call = std::dynamic_pointer_cast<const Call>(eval->expr_);
-    }
-    auto op = call ? std::dynamic_pointer_cast<const Op>(call->op_) : nullptr;
-    if (op) {
-      const std::string& op_name = op->name_;
-      if (op_name == "system.reserve_buffer") {
-        metadata.has_reserve_buffer = true;
-      } else if (op_name == "system.import_peer_buffer") {
-        metadata.has_import_peer_buffer = true;
-      } else if (op_name == "system.aic_initialize_pipe") {
-        metadata.has_aic_initialize_pipe = true;
-      } else if (op_name == "system.aiv_initialize_pipe") {
-        metadata.has_aiv_initialize_pipe = true;
-      }
-    }
-
-    CrossCorePipeMetadata stmt_metadata;
-    CollectCrossCorePipeMetadata({stmt}, stmt_metadata);
-    if (stmt_metadata.HasCrossCoreOps()) {
-      break;
-    }
-  }
-  return metadata;
-}
-
-struct AutomaticPipeSetup {
-  std::vector<StmtPtr> aic_stmts;
-  std::vector<StmtPtr> aiv_stmts;
-};
-
-AutomaticPipeSetup BuildAutomaticPipeSetup(const std::string& func_name, const std::string& aic_name,
-                                           const std::string& aiv_name, const std::vector<StmtPtr>& aic_stmts,
-                                           const std::vector<StmtPtr>& aiv_stmts, const Span& span) {
-  CrossCorePipeMetadata aic_metadata;
-  CollectCrossCorePipeMetadata(aic_stmts, aic_metadata);
-  CrossCorePipeMetadata aiv_metadata;
-  CollectCrossCorePipeMetadata(aiv_stmts, aiv_metadata);
-  CrossCorePipeMetadata combined = MergeCrossCorePipeMetadata(aic_metadata, aiv_metadata);
-
-  if (!combined.HasCrossCoreOps() || aic_metadata.HasAnySetup() || aiv_metadata.HasAnySetup()) {
-    return {};
-  }
-
-  const int dir_mask = BuildDirMask(combined);
-  auto common_slot_size = GetCommonSlotSizeBytes(combined);
-  if (dir_mask == 0 || !common_slot_size.has_value()) {
-    return {};
-  }
-
-  const int64_t buffer_size = common_slot_size.value() * GetSlotNumForDirMask(dir_mask);
-  CHECK(common_slot_size.value() <= std::numeric_limits<int>::max())
-      << "Cross-core slot_size out of range: " << common_slot_size.value();
-  const int slot_size_bytes = static_cast<int>(common_slot_size.value());
-  AutomaticPipeSetup setup;
-
-  if (dir_mask & kDirMaskV2C) {
-    setup.aic_stmts.push_back(std::make_shared<AssignStmt>(
-        std::make_shared<Var>(BuildPipeBufferName(func_name, PipeDirection::V2C), GetUnknownType(), span),
-        CreateReserveBuffer(BuildPipeBufferName(func_name, PipeDirection::V2C), buffer_size, span), span));
-    setup.aiv_stmts.push_back(std::make_shared<AssignStmt>(
-        std::make_shared<Var>(BuildPipeBufferName(func_name, PipeDirection::V2C) + "_import",
-                              GetUnknownType(), span),
-        CreateImportPeerBuffer(BuildPipeBufferName(func_name, PipeDirection::V2C), aic_name, span), span));
-  }
-
-  if (dir_mask & kDirMaskC2V) {
-    setup.aiv_stmts.push_back(std::make_shared<AssignStmt>(
-        std::make_shared<Var>(BuildPipeBufferName(func_name, PipeDirection::C2V), GetUnknownType(), span),
-        CreateReserveBuffer(BuildPipeBufferName(func_name, PipeDirection::C2V), buffer_size, span), span));
-    setup.aic_stmts.push_back(std::make_shared<AssignStmt>(
-        std::make_shared<Var>(BuildPipeBufferName(func_name, PipeDirection::C2V) + "_import",
-                              GetUnknownType(), span),
-        CreateImportPeerBuffer(BuildPipeBufferName(func_name, PipeDirection::C2V), aiv_name, span), span));
-  }
-
-  setup.aic_stmts.push_back(
-      std::make_shared<EvalStmt>(CreateInitializePipe(CoreSide::AIC, dir_mask, slot_size_bytes, span), span));
-  setup.aiv_stmts.push_back(
-      std::make_shared<EvalStmt>(CreateInitializePipe(CoreSide::AIV, dir_mask, slot_size_bytes, span), span));
-
-  return setup;
-}
-
-std::vector<StmtPtr> PrependPipeSetup(const std::vector<StmtPtr>& prologue,
-                                      const std::vector<StmtPtr>& body) {
-  if (prologue.empty()) return body;
-  std::vector<StmtPtr> result;
-  result.reserve(prologue.size() + body.size());
-  result.insert(result.end(), prologue.begin(), prologue.end());
-  result.insert(result.end(), body.begin(), body.end());
-  return result;
-}
 
 // ============================================================================
 // Recursive Affinity Analysis
@@ -613,17 +181,6 @@ CoreAffinity AnalyzeStmtAffinity(const StmtPtr& stmt, std::unordered_map<const S
 // CV Boundary Move Collection
 // ============================================================================
 
-/// Information about a single CV boundary tile.move found in the IR.
-struct CVBoundaryMove {
-  CVDirection direction;
-  VarPtr dest_var;      // AssignStmt LHS
-  ExprPtr source_tile;  // First arg of tile.move
-  TypePtr result_type;  // Return type of the tile.move call
-  /// Original tpop Call that defines source_tile (nullptr if source is not a tpop result).
-  /// Used to propagate kwargs (e.g., split) from the user-written tpop to the replacement tpop.
-  CallPtr source_tpop_call;
-};
-
 /// Collect all CV boundary tile.move statements recursively.
 /// Also looks up whether the source tile is defined by a tpop call, so that kwargs
 /// (e.g., split) can be propagated to the replacement tpop.
@@ -691,632 +248,11 @@ CallPtr CreateTpop(const std::string& op_name, const TypePtr& tile_type, const S
                                 CleanTileType(tile_type), span);
 }
 
-std::string GetTfreeOpName(CoreSide side) {
-  return (side == CoreSide::AIC) ? "system.tfree_to_aiv" : "system.tfree_to_aic";
-}
-
-CallPtr CreateTfree(CoreSide side, const ExprPtr& tile, const Span& span) {
-  return OpRegistry::GetInstance().Create(GetTfreeOpName(side), {tile}, {}, span);
-}
-
 CallPtr CreateMove(const ExprPtr& tile, MemorySpace target_memory, const TypePtr& result_type,
                    const Span& span) {
   auto op = OpRegistry::GetInstance().GetOp("tile.move");
   std::vector<std::pair<std::string, std::any>> kwargs{{"target_memory", std::any(target_memory)}};
   return std::make_shared<Call>(op, std::vector<ExprPtr>{tile}, std::move(kwargs), result_type, span);
-}
-
-// ============================================================================
-// Recursive Dead Code Elimination
-// ============================================================================
-
-/// Extract the Op name from an AssignStmt or EvalStmt containing a Call.
-/// Returns empty string if the statement doesn't match this pattern.
-std::string GetStmtOpName(const StmtPtr& stmt) {
-  CallPtr call;
-  if (auto assign = std::dynamic_pointer_cast<const AssignStmt>(stmt)) {
-    call = std::dynamic_pointer_cast<const Call>(assign->value_);
-  } else if (auto eval = std::dynamic_pointer_cast<const EvalStmt>(stmt)) {
-    call = std::dynamic_pointer_cast<const Call>(eval->expr_);
-  }
-  if (call && call->op_) {
-    if (auto op = std::dynamic_pointer_cast<const Op>(call->op_)) {
-      return op->name_;
-    }
-  }
-  return "";
-}
-
-bool IsSideEffectOp(const StmtPtr& stmt) {
-  static const std::unordered_set<std::string> side_effect_ops = {"tile.tpush_to_aiv",
-                                                                  "tile.tpush_to_aic",
-                                                                  "tile.tpop_from_aic",
-                                                                  "tile.tpop_from_aiv",
-                                                                  "tile.store",
-                                                                  "tile.assemble",
-                                                                  "system.tfree_to_aic",
-                                                                  "system.tfree_to_aiv",
-                                                                  "system.reserve_buffer",
-                                                                  "system.import_peer_buffer",
-                                                                  "system.aic_initialize_pipe",
-                                                                  "system.aiv_initialize_pipe"};
-  return side_effect_ops.count(GetStmtOpName(stmt)) > 0;
-}
-
-void CollectAllAssignStmts(const std::vector<StmtPtr>& stmts,
-                           std::vector<std::shared_ptr<const AssignStmt>>& assigns) {
-  for (const auto& stmt : stmts) {
-    if (auto assign = std::dynamic_pointer_cast<const AssignStmt>(stmt)) {
-      assigns.push_back(assign);
-    }
-    if (auto for_stmt = std::dynamic_pointer_cast<const ForStmt>(stmt)) {
-      CollectAllAssignStmts(FlattenBody(for_stmt->body_), assigns);
-    } else if (auto if_stmt = std::dynamic_pointer_cast<const IfStmt>(stmt)) {
-      CollectAllAssignStmts(FlattenBody(if_stmt->then_body_), assigns);
-      if (if_stmt->else_body_.has_value()) {
-        CollectAllAssignStmts(FlattenBody(if_stmt->else_body_.value()), assigns);
-      }
-    } else if (auto while_stmt = std::dynamic_pointer_cast<const WhileStmt>(stmt)) {
-      CollectAllAssignStmts(FlattenBody(while_stmt->body_), assigns);
-    }
-  }
-}
-
-void FindLiveRootsRecursive(const std::vector<StmtPtr>& stmts, std::unordered_set<const Var*>& live) {
-  for (const auto& stmt : stmts) {
-    if (std::dynamic_pointer_cast<const ReturnStmt>(stmt) ||
-        std::dynamic_pointer_cast<const YieldStmt>(stmt) || IsSideEffectOp(stmt)) {
-      outline_utils::VarRefCollector refs;
-      refs.VisitStmt(stmt);
-      live.insert(refs.var_refs.begin(), refs.var_refs.end());
-      // Mark LHS of side-effect assignments as live for downstream propagation
-      if (auto assign = std::dynamic_pointer_cast<const AssignStmt>(stmt)) {
-        live.insert(assign->var_.get());
-      }
-    }
-    // Collect variable refs from control expressions and iter_args init values
-    auto collect_iter_arg_refs = [&](const auto& loop_stmt) {
-      for (const auto& iter_arg : loop_stmt->iter_args_) {
-        outline_utils::VarRefCollector refs;
-        refs.VisitExpr(iter_arg->initValue_);
-        live.insert(refs.var_refs.begin(), refs.var_refs.end());
-      }
-    };
-    auto collect_expr_refs = [&](const ExprPtr& expr) {
-      outline_utils::VarRefCollector refs;
-      refs.VisitExpr(expr);
-      live.insert(refs.var_refs.begin(), refs.var_refs.end());
-    };
-
-    if (auto for_stmt = std::dynamic_pointer_cast<const ForStmt>(stmt)) {
-      collect_expr_refs(for_stmt->start_);
-      collect_expr_refs(for_stmt->stop_);
-      collect_expr_refs(for_stmt->step_);
-      collect_iter_arg_refs(for_stmt);
-      FindLiveRootsRecursive(FlattenBody(for_stmt->body_), live);
-    } else if (auto if_stmt = std::dynamic_pointer_cast<const IfStmt>(stmt)) {
-      collect_expr_refs(if_stmt->condition_);
-      FindLiveRootsRecursive(FlattenBody(if_stmt->then_body_), live);
-      if (if_stmt->else_body_.has_value()) {
-        FindLiveRootsRecursive(FlattenBody(if_stmt->else_body_.value()), live);
-      }
-    } else if (auto while_stmt = std::dynamic_pointer_cast<const WhileStmt>(stmt)) {
-      collect_expr_refs(while_stmt->condition_);
-      collect_iter_arg_refs(while_stmt);
-      FindLiveRootsRecursive(FlattenBody(while_stmt->body_), live);
-    }
-  }
-}
-
-std::vector<StmtPtr> FilterDeadCode(const std::vector<StmtPtr>& stmts,
-                                    const std::unordered_set<const Var*>& live) {
-  std::vector<StmtPtr> result;
-  for (const auto& stmt : stmts) {
-    if (auto assign = std::dynamic_pointer_cast<const AssignStmt>(stmt)) {
-      if (live.count(assign->var_.get()) || IsSideEffectOp(stmt)) {
-        result.push_back(stmt);
-      }
-    } else if (auto for_stmt = std::dynamic_pointer_cast<const ForStmt>(stmt)) {
-      auto filtered = FilterDeadCode(FlattenBody(for_stmt->body_), live);
-      result.push_back(std::make_shared<ForStmt>(
-          for_stmt->loop_var_, for_stmt->start_, for_stmt->stop_, for_stmt->step_, for_stmt->iter_args_,
-          MakeBody(filtered, for_stmt->span_), for_stmt->return_vars_, for_stmt->span_, for_stmt->kind_,
-          for_stmt->chunk_size_, for_stmt->chunk_policy_, for_stmt->loop_origin_));
-    } else if (auto if_stmt = std::dynamic_pointer_cast<const IfStmt>(stmt)) {
-      auto filtered_then = FilterDeadCode(FlattenBody(if_stmt->then_body_), live);
-      std::optional<StmtPtr> filtered_else;
-      if (if_stmt->else_body_.has_value()) {
-        auto fe = FilterDeadCode(FlattenBody(if_stmt->else_body_.value()), live);
-        filtered_else = MakeBody(fe, if_stmt->span_);
-      }
-      result.push_back(std::make_shared<IfStmt>(if_stmt->condition_, MakeBody(filtered_then, if_stmt->span_),
-                                                filtered_else, if_stmt->return_vars_, if_stmt->span_));
-    } else if (auto while_stmt = std::dynamic_pointer_cast<const WhileStmt>(stmt)) {
-      auto filtered = FilterDeadCode(FlattenBody(while_stmt->body_), live);
-      result.push_back(std::make_shared<WhileStmt>(while_stmt->condition_, while_stmt->iter_args_,
-                                                   MakeBody(filtered, while_stmt->span_),
-                                                   while_stmt->return_vars_, while_stmt->span_));
-    } else {
-      // ReturnStmt, EvalStmt (side-effect), etc. — always keep
-      result.push_back(stmt);
-    }
-  }
-  return result;
-}
-
-std::vector<StmtPtr> EliminateDeadCode(const std::vector<StmtPtr>& stmts) {
-  std::unordered_set<const Var*> live;
-
-  // Find initial live set from returns and side-effect ops at all nesting levels
-  FindLiveRootsRecursive(stmts, live);
-
-  // Collect all assignments at all nesting levels for backward propagation
-  std::vector<std::shared_ptr<const AssignStmt>> all_assigns;
-  CollectAllAssignStmts(stmts, all_assigns);
-
-  // Backward pass: propagate liveness
-  bool changed = true;
-  while (changed) {
-    changed = false;
-    for (auto it = all_assigns.rbegin(); it != all_assigns.rend(); ++it) {
-      if (!live.count((*it)->var_.get())) continue;
-
-      outline_utils::VarRefCollector refs;
-      refs.VisitExpr((*it)->value_);
-      for (const Var* ref : refs.var_refs) {
-        if (!live.count(ref)) {
-          live.insert(ref);
-          changed = true;
-        }
-      }
-    }
-  }
-
-  return FilterDeadCode(stmts, live);
-}
-
-// ============================================================================
-// Loop Iter_arg Cleanup (strip dead, fix dangling references)
-// ============================================================================
-
-// --- Reconstruction helpers (reduce repetitive loop/if rebuilds) ---
-
-/// Rebuild a ForStmt with a new body, preserving all other fields.
-StmtPtr RebuildForStmt(const std::shared_ptr<const ForStmt>& f, const StmtPtr& new_body) {
-  return std::make_shared<ForStmt>(f->loop_var_, f->start_, f->stop_, f->step_, f->iter_args_, new_body,
-                                   f->return_vars_, f->span_, f->kind_, f->chunk_size_, f->chunk_policy_,
-                                   f->loop_origin_);
-}
-
-/// Rebuild a ForStmt with new iter_args, body, and return_vars.
-StmtPtr RebuildForStmt(const std::shared_ptr<const ForStmt>& f, const std::vector<IterArgPtr>& iter_args,
-                       const StmtPtr& new_body, const std::vector<VarPtr>& return_vars) {
-  return std::make_shared<ForStmt>(f->loop_var_, f->start_, f->stop_, f->step_, iter_args, new_body,
-                                   return_vars, f->span_, f->kind_, f->chunk_size_, f->chunk_policy_,
-                                   f->loop_origin_);
-}
-
-/// Rebuild a WhileStmt with a new body, preserving all other fields.
-StmtPtr RebuildWhileStmt(const std::shared_ptr<const WhileStmt>& w, const StmtPtr& new_body) {
-  return std::make_shared<WhileStmt>(w->condition_, w->iter_args_, new_body, w->return_vars_, w->span_);
-}
-
-/// Rebuild a WhileStmt with new iter_args, body, and return_vars.
-StmtPtr RebuildWhileStmt(const std::shared_ptr<const WhileStmt>& w, const std::vector<IterArgPtr>& iter_args,
-                         const StmtPtr& new_body, const std::vector<VarPtr>& return_vars) {
-  return std::make_shared<WhileStmt>(w->condition_, iter_args, new_body, return_vars, w->span_);
-}
-
-/// Rebuild an IfStmt with new then/else bodies, preserving condition and return_vars.
-StmtPtr RebuildIfStmt(const std::shared_ptr<const IfStmt>& s, const std::vector<StmtPtr>& new_then,
-                      const std::optional<std::vector<StmtPtr>>& new_else_stmts) {
-  std::optional<StmtPtr> new_else;
-  if (new_else_stmts.has_value()) {
-    new_else = MakeBody(new_else_stmts.value(), s->span_);
-  }
-  return std::make_shared<IfStmt>(s->condition_, MakeBody(new_then, s->span_), new_else, s->return_vars_,
-                                  s->span_);
-}
-
-/// Process an IfStmt's else branch: flatten, apply transform, return processed stmts.
-/// Returns nullopt if the IfStmt has no else branch.
-template <typename Fn>
-std::optional<std::vector<StmtPtr>> ProcessElseBranch(const std::shared_ptr<const IfStmt>& if_stmt,
-                                                      Fn&& transform) {
-  if (!if_stmt->else_body_.has_value()) return std::nullopt;
-  return transform(FlattenBody(if_stmt->else_body_.value()));
-}
-
-/// Apply a function to the last statement of a compound stmt tree (IfStmt/SeqStmts).
-/// The transform function may return nullptr to signal removal of the last statement.
-/// Used by FilterYieldStmt (nullable) and FixDanglingYieldStmt (always non-null).
-template <typename Fn>
-StmtPtr TransformLastStmt(const StmtPtr& stmt, Fn&& transform) {
-  // Helper: apply transform to the back of a statement list, removing if null
-  auto apply_to_back = [&](std::vector<StmtPtr>& stmts) {
-    if (stmts.empty()) return;
-    auto result = TransformLastStmt(stmts.back(), transform);
-    if (result) {
-      stmts.back() = result;
-    } else {
-      stmts.pop_back();
-    }
-  };
-
-  if (auto if_stmt = std::dynamic_pointer_cast<const IfStmt>(stmt)) {
-    auto then_stmts = FlattenBody(if_stmt->then_body_);
-    apply_to_back(then_stmts);
-    auto else_stmts = ProcessElseBranch(if_stmt, [&](std::vector<StmtPtr> es) {
-      apply_to_back(es);
-      return es;
-    });
-    return RebuildIfStmt(if_stmt, then_stmts, else_stmts);
-  }
-
-  if (auto seq = std::dynamic_pointer_cast<const SeqStmts>(stmt)) {
-    auto seq_stmts = seq->stmts_;
-    apply_to_back(seq_stmts);
-    return MakeBody(seq_stmts, seq->span_);
-  }
-
-  return transform(stmt);
-}
-
-// --- Variable reference collection helpers ---
-
-/// Collect variable references from body statements, skipping top-level YieldStmt.
-/// This determines which iter_arg variables are actually used by computation
-/// statements, excluding the yield that merely carries values between iterations.
-/// NOTE: Yields nested inside IfStmt/SeqStmts are still visited (conservative --
-/// may keep iter_args alive that are only referenced in conditional yields).
-/// This is safe: over-conservative keeps correctness, and such patterns are rare.
-void CollectBodyRefsSkippingYield(const std::vector<StmtPtr>& stmts, std::unordered_set<const Var*>& refs) {
-  for (const auto& stmt : stmts) {
-    if (std::dynamic_pointer_cast<const YieldStmt>(stmt)) continue;
-    outline_utils::VarRefCollector collector;
-    collector.VisitStmt(stmt);
-    refs.insert(collector.var_refs.begin(), collector.var_refs.end());
-  }
-}
-
-/// Recursively filter YieldStmt values by kept_indices.
-/// Returns nullptr if all yield values are stripped and the yield should be removed.
-StmtPtr FilterYieldStmt(const StmtPtr& stmt, const std::vector<size_t>& kept_indices) {
-  return TransformLastStmt(stmt, [&](const StmtPtr& s) -> StmtPtr {
-    auto yield_stmt = std::dynamic_pointer_cast<const YieldStmt>(s);
-    if (!yield_stmt) return s;
-    if (kept_indices.empty()) return nullptr;
-    std::vector<ExprPtr> new_values;
-    for (size_t idx : kept_indices) {
-      INTERNAL_CHECK(idx < yield_stmt->value_.size())
-          << "Internal error: yield index " << idx << " out of range " << yield_stmt->value_.size();
-      new_values.push_back(yield_stmt->value_[idx]);
-    }
-    return std::make_shared<YieldStmt>(new_values, yield_stmt->span_);
-  });
-}
-
-/// Rebuild a For or While loop with the given iter_args, body, and return_vars.
-StmtPtr RebuildLoop(const std::shared_ptr<const ForStmt>& for_stmt,
-                    const std::shared_ptr<const WhileStmt>& while_stmt,
-                    const std::vector<IterArgPtr>& iter_args, const StmtPtr& new_body,
-                    const std::vector<VarPtr>& return_vars) {
-  if (for_stmt) return RebuildForStmt(for_stmt, iter_args, new_body, return_vars);
-  return RebuildWhileStmt(while_stmt, iter_args, new_body, return_vars);
-}
-
-/// Strip dead iter_args from ForStmt/WhileStmt in the given statement list.
-/// An iter_arg is dead if:
-///   1. Its variable name is not referenced by any non-yield statement in the loop body, AND
-///   2. Its corresponding return_var is not referenced by any statement after the loop.
-/// Dead iter_args, their return_vars, and corresponding yield values are removed.
-std::vector<StmtPtr> StripDeadIterArgs(const std::vector<StmtPtr>& stmts) {
-  // Precompute suffix reference sets (back-to-front) to avoid O(N^2) re-scanning.
-  // suffix_refs[i] contains all variable references from stmts[i+1..end].
-  std::vector<std::unordered_set<const Var*>> suffix_refs(stmts.size());
-  for (size_t i = stmts.size(); i-- > 0;) {
-    if (i + 1 < stmts.size()) {
-      suffix_refs[i] = suffix_refs[i + 1];
-    }
-    outline_utils::VarRefCollector collector;
-    collector.VisitStmt(stmts[i]);
-    suffix_refs[i].insert(collector.var_refs.begin(), collector.var_refs.end());
-  }
-
-  std::vector<StmtPtr> result;
-
-  for (size_t idx = 0; idx < stmts.size(); ++idx) {
-    auto for_stmt = std::dynamic_pointer_cast<const ForStmt>(stmts[idx]);
-    auto while_stmt = std::dynamic_pointer_cast<const WhileStmt>(stmts[idx]);
-
-    // For non-loop statements, recurse into compound children
-    if (!for_stmt && !while_stmt) {
-      if (auto if_stmt = std::dynamic_pointer_cast<const IfStmt>(stmts[idx])) {
-        auto new_then = StripDeadIterArgs(FlattenBody(if_stmt->then_body_));
-        auto new_else =
-            ProcessElseBranch(if_stmt, [](const std::vector<StmtPtr>& es) { return StripDeadIterArgs(es); });
-        result.push_back(RebuildIfStmt(if_stmt, new_then, new_else));
-      } else {
-        result.push_back(stmts[idx]);
-      }
-      continue;
-    }
-
-    const auto& iter_args = for_stmt ? for_stmt->iter_args_ : while_stmt->iter_args_;
-    const auto& return_vars = for_stmt ? for_stmt->return_vars_ : while_stmt->return_vars_;
-    const auto& body = for_stmt ? for_stmt->body_ : while_stmt->body_;
-    const auto& span = for_stmt ? for_stmt->span_ : while_stmt->span_;
-
-    // Recursively process the loop body first (bottom-up)
-    auto processed_body = StripDeadIterArgs(FlattenBody(body));
-
-    // No iter_args — just rebuild with processed body
-    if (iter_args.empty()) {
-      result.push_back(
-          RebuildLoop(for_stmt, while_stmt, iter_args, MakeBody(processed_body, span), return_vars));
-      continue;
-    }
-
-    // Collect var refs from processed loop body, excluding top-level YieldStmt
-    std::unordered_set<const Var*> body_refs;
-    CollectBodyRefsSkippingYield(processed_body, body_refs);
-
-    // O(1) lookup into precomputed suffix refs for statements after this loop
-    static const std::unordered_set<const Var*> kEmptyRefs;
-    const auto& after_refs = (idx + 1 < stmts.size()) ? suffix_refs[idx + 1] : kEmptyRefs;
-
-    // Determine which iter_args are live
-    std::vector<size_t> kept_indices;
-    for (size_t i = 0; i < iter_args.size(); ++i) {
-      bool used_in_body = body_refs.count(iter_args[i].get()) > 0;
-      bool return_var_used = i < return_vars.size() && after_refs.count(return_vars[i].get()) > 0;
-      if (used_in_body || return_var_used) {
-        kept_indices.push_back(i);
-      }
-    }
-
-    // Filter iter_args and return_vars by kept_indices (identity when all alive)
-    std::vector<IterArgPtr> new_iter_args;
-    std::vector<VarPtr> new_return_vars;
-    for (size_t i : kept_indices) {
-      new_iter_args.push_back(iter_args[i]);
-      if (i < return_vars.size()) {
-        new_return_vars.push_back(return_vars[i]);
-      }
-    }
-
-    // Filter YieldStmt values when some iter_args were stripped
-    if (kept_indices.size() < iter_args.size() && !processed_body.empty()) {
-      auto filtered_last = FilterYieldStmt(processed_body.back(), kept_indices);
-      if (filtered_last) {
-        processed_body.back() = filtered_last;
-      } else {
-        processed_body.pop_back();
-      }
-    }
-
-    result.push_back(
-        RebuildLoop(for_stmt, while_stmt, new_iter_args, MakeBody(processed_body, span), new_return_vars));
-  }
-
-  return result;
-}
-
-/// Build a map from variable name to its defining AssignStmt in the original body.
-/// Only collects top-level assignments (init values are typically at the top level).
-void BuildDefMap(const std::vector<StmtPtr>& stmts, std::unordered_map<const Var*, StmtPtr>& def_map) {
-  for (const auto& stmt : stmts) {
-    if (auto assign = std::dynamic_pointer_cast<const AssignStmt>(stmt)) {
-      def_map[assign->var_.get()] = stmt;
-    }
-  }
-}
-
-/// Recursively collect the definition chain for a variable from the def_map.
-/// Returns definitions in dependency order (dependencies first).
-void PullDefinitionChain(const Var* var_ptr, const std::unordered_map<const Var*, StmtPtr>& def_map,
-                         const std::unordered_set<const Var*>& already_defined,
-                         std::unordered_set<const Var*>& pulled, std::vector<StmtPtr>& out) {
-  if (pulled.count(var_ptr) || already_defined.count(var_ptr)) return;
-  auto it = def_map.find(var_ptr);
-  if (it == def_map.end()) return;
-
-  pulled.insert(var_ptr);
-
-  // Recursively pull dependencies first
-  auto assign = std::dynamic_pointer_cast<const AssignStmt>(it->second);
-  if (assign) {
-    outline_utils::VarRefCollector refs;
-    refs.VisitExpr(assign->value_);
-    for (const Var* dep : refs.var_refs) {
-      PullDefinitionChain(dep, def_map, already_defined, pulled, out);
-    }
-  }
-
-  out.push_back(it->second);
-}
-
-/// Fix alive iter_args whose init values reference undefined variables.
-/// Pulls the missing definitions from the original (pre-split) body.
-/// Uses a prefix-only `defined_so_far` set (variables defined before the current
-/// statement) to avoid treating non-dominating definitions as available.
-std::vector<StmtPtr> FixupIterArgInitValues(const std::vector<StmtPtr>& stmts,
-                                            const std::unordered_map<const Var*, StmtPtr>& original_def_map) {
-  auto recurse = [&](const std::vector<StmtPtr>& s) { return FixupIterArgInitValues(s, original_def_map); };
-
-  // Build prefix-only defined set: track definitions as we scan statements in order.
-  std::unordered_set<const Var*> defined_so_far;
-  std::vector<StmtPtr> result;
-  std::unordered_set<const Var*> pulled;
-
-  for (const auto& stmt : stmts) {
-    auto for_stmt = std::dynamic_pointer_cast<const ForStmt>(stmt);
-    auto while_stmt = std::dynamic_pointer_cast<const WhileStmt>(stmt);
-
-    // Pull missing init value definitions for loop iter_args
-    const std::vector<IterArgPtr>* iter_args_ptr = nullptr;
-    if (for_stmt) {
-      iter_args_ptr = &for_stmt->iter_args_;
-    } else if (while_stmt) {
-      iter_args_ptr = &while_stmt->iter_args_;
-    }
-    if (iter_args_ptr && !iter_args_ptr->empty()) {
-      std::vector<StmtPtr> missing_defs;
-      for (const auto& iter_arg : *iter_args_ptr) {
-        outline_utils::VarRefCollector refs;
-        refs.VisitExpr(iter_arg->initValue_);
-        for (const Var* ref : refs.var_refs) {
-          if (!defined_so_far.count(ref) && !pulled.count(ref)) {
-            PullDefinitionChain(ref, original_def_map, defined_so_far, pulled, missing_defs);
-          }
-        }
-      }
-      // Add pulled definitions to defined_so_far so later statements see them
-      for (const auto& def : missing_defs) {
-        if (auto assign = std::dynamic_pointer_cast<const AssignStmt>(def)) {
-          defined_so_far.insert(assign->var_.get());
-        }
-      }
-      result.insert(result.end(), missing_defs.begin(), missing_defs.end());
-    }
-
-    // Track definitions from the current statement before recursing
-    outline_utils::VarDefCollector stmt_defs;
-    stmt_defs.VisitStmt(stmt);
-    defined_so_far.insert(stmt_defs.var_defs.begin(), stmt_defs.var_defs.end());
-
-    // Recurse into compound statements
-    if (for_stmt) {
-      auto new_body = recurse(FlattenBody(for_stmt->body_));
-      result.push_back(RebuildForStmt(for_stmt, MakeBody(new_body, for_stmt->span_)));
-    } else if (auto if_stmt = std::dynamic_pointer_cast<const IfStmt>(stmt)) {
-      auto new_then = recurse(FlattenBody(if_stmt->then_body_));
-      auto new_else = ProcessElseBranch(if_stmt, [&](const std::vector<StmtPtr>& es) { return recurse(es); });
-      result.push_back(RebuildIfStmt(if_stmt, new_then, new_else));
-    } else if (while_stmt) {
-      auto new_body = recurse(FlattenBody(while_stmt->body_));
-      result.push_back(RebuildWhileStmt(while_stmt, MakeBody(new_body, while_stmt->span_)));
-    } else {
-      result.push_back(stmt);
-    }
-  }
-
-  return result;
-}
-
-/// Recursively fix dangling YieldStmt values by replacing them with the iter_arg
-/// (identity yield -- preserves previous iteration's value on this core side).
-StmtPtr FixDanglingYieldStmt(const StmtPtr& stmt, const std::vector<IterArgPtr>& iter_args,
-                             const std::unordered_set<const Var*>& defined_vars) {
-  return TransformLastStmt(stmt, [&](const StmtPtr& s) -> StmtPtr {
-    auto yield_stmt = std::dynamic_pointer_cast<const YieldStmt>(s);
-    if (!yield_stmt) return s;
-
-    std::vector<ExprPtr> new_values;
-    for (size_t i = 0; i < yield_stmt->value_.size(); ++i) {
-      outline_utils::VarRefCollector refs;
-      refs.VisitExpr(yield_stmt->value_[i]);
-      bool has_undefined = std::any_of(refs.var_refs.begin(), refs.var_refs.end(),
-                                       [&](const Var* ref) { return !defined_vars.count(ref); });
-      if (has_undefined && i < iter_args.size()) {
-        new_values.push_back(iter_args[i]);
-      } else {
-        new_values.push_back(yield_stmt->value_[i]);
-      }
-    }
-    return std::make_shared<YieldStmt>(new_values, yield_stmt->span_);
-  });
-}
-
-/// Fix dangling yields that appear anywhere in a loop body statement list.
-/// This covers both the loop's final YieldStmt and nested IfStmt branch yields
-/// that carry loop state forward through a conditional.
-std::vector<StmtPtr> FixDanglingLoopBodyYields(const std::vector<StmtPtr>& stmts,
-                                               const std::vector<IterArgPtr>& iter_args,
-                                               const std::unordered_set<const Var*>& defined_vars) {
-  std::vector<StmtPtr> result;
-  result.reserve(stmts.size());
-  for (const auto& stmt : stmts) {
-    result.push_back(FixDanglingYieldStmt(stmt, iter_args, defined_vars));
-  }
-  return result;
-}
-
-/// Fix dangling yield values in ForStmt/WhileStmt bodies.
-/// When a yield value references an undefined variable (its definition was stripped
-/// during core body splitting), replace it with the corresponding iter_arg variable
-/// (identity yield -- preserves the value from the previous iteration on this core side).
-/// Uses a prefix-only `defined_so_far` set to avoid treating non-dominating definitions
-/// (defined later in the same scope) as available.
-std::vector<StmtPtr> FixupDanglingYieldValues(const std::vector<StmtPtr>& stmts) {
-  // Build prefix-only defined set: track definitions as we scan statements in order.
-  std::unordered_set<const Var*> defined_so_far;
-
-  std::vector<StmtPtr> result;
-  for (const auto& stmt : stmts) {
-    auto for_stmt = std::dynamic_pointer_cast<const ForStmt>(stmt);
-    auto while_stmt = std::dynamic_pointer_cast<const WhileStmt>(stmt);
-
-    if ((for_stmt && !for_stmt->iter_args_.empty()) || (while_stmt && !while_stmt->iter_args_.empty())) {
-      const auto& iter_args = for_stmt ? for_stmt->iter_args_ : while_stmt->iter_args_;
-      const auto& body = for_stmt ? for_stmt->body_ : while_stmt->body_;
-
-      // Collect defined vars within the loop body + preceding scope
-      outline_utils::VarDefCollector body_def_collector;
-      body_def_collector.VisitStmt(body);
-      auto all_defined = defined_so_far;
-      all_defined.insert(body_def_collector.var_defs.begin(), body_def_collector.var_defs.end());
-
-      // Recursively process body, then fix dangling yields in the loop tail and
-      // any nested conditional branches carrying loop state.
-      auto body_stmts = FixupDanglingYieldValues(FlattenBody(body));
-      body_stmts = FixDanglingLoopBodyYields(body_stmts, iter_args, all_defined);
-
-      const auto& span = for_stmt ? for_stmt->span_ : while_stmt->span_;
-      if (for_stmt) {
-        result.push_back(RebuildForStmt(for_stmt, MakeBody(body_stmts, span)));
-      } else {
-        result.push_back(RebuildWhileStmt(while_stmt, MakeBody(body_stmts, span)));
-      }
-    } else if (auto if_stmt = std::dynamic_pointer_cast<const IfStmt>(stmt)) {
-      auto new_then = FixupDanglingYieldValues(FlattenBody(if_stmt->then_body_));
-      auto new_else = ProcessElseBranch(
-          if_stmt, [](const std::vector<StmtPtr>& es) { return FixupDanglingYieldValues(es); });
-      result.push_back(RebuildIfStmt(if_stmt, new_then, new_else));
-    } else {
-      result.push_back(stmt);
-    }
-
-    // Track definitions from the current statement for subsequent iterations
-    outline_utils::VarDefCollector stmt_defs;
-    stmt_defs.VisitStmt(stmt);
-    defined_so_far.insert(stmt_defs.var_defs.begin(), stmt_defs.var_defs.end());
-  }
-
-  return result;
-}
-
-/// Final cleanup after one core side has been split out of the mixed body.
-///
-/// The split step intentionally preserves SHARED control-flow structure on both
-/// sides. That can leave loop-carried state temporarily inconsistent until we
-/// repair it in order:
-///   1. Strip iter_args whose carried values are dead on this side.
-///   2. Pull back any now-missing init-value definitions for surviving iter_args.
-///   3. Rewrite dangling yields to identity yields when a branch-local value was
-///      stripped on this side.
-///   4. Run DCE once the loop state is structurally valid again.
-///   5. Normalize loop-carried state one more time, because DCE may remove a
-///      SHARED-only post-loop use that temporarily kept an iter_arg alive.
-///   6. Run DCE again to clean up init-value chains exposed by the second strip.
-std::vector<StmtPtr> FinalizeSplitCoreBody(const std::vector<StmtPtr>& stmts,
-                                           const std::unordered_map<const Var*, StmtPtr>& original_def_map) {
-  auto repaired = StripDeadIterArgs(stmts);
-  repaired = FixupIterArgInitValues(repaired, original_def_map);
-  repaired = FixupDanglingYieldValues(repaired);
-  repaired = EliminateDeadCode(repaired);
-  repaired = StripDeadIterArgs(repaired);
-  return EliminateDeadCode(repaired);
 }
 
 // ============================================================================
@@ -1462,306 +398,6 @@ std::vector<StmtPtr> BuildCoreBody(CoreSide side, const std::vector<StmtPtr>& st
   }
 
   return result;
-}
-
-bool IsTpopAssignStmt(const StmtPtr& stmt, VarPtr* result_var = nullptr) {
-  auto assign = std::dynamic_pointer_cast<const AssignStmt>(stmt);
-  if (!assign) return false;
-  auto call = std::dynamic_pointer_cast<const Call>(assign->value_);
-  auto op = call ? std::dynamic_pointer_cast<const Op>(call->op_) : nullptr;
-  if (!op || (op->name_ != "tile.tpop_from_aiv" && op->name_ != "tile.tpop_from_aic")) {
-    return false;
-  }
-  if (result_var) *result_var = assign->var_;
-  return true;
-}
-
-bool IsExpectedTpopOp(const std::string& op_name, FunctionType func_type) {
-  if (func_type == FunctionType::AIC) return op_name == "tile.tpop_from_aiv";
-  if (func_type == FunctionType::AIV) return op_name == "tile.tpop_from_aic";
-  return false;
-}
-
-bool IsExpectedTpopAssignStmt(const StmtPtr& stmt, FunctionType func_type, VarPtr* result_var = nullptr) {
-  auto assign = std::dynamic_pointer_cast<const AssignStmt>(stmt);
-  if (!assign) return false;
-  auto call = std::dynamic_pointer_cast<const Call>(assign->value_);
-  auto op = call ? std::dynamic_pointer_cast<const Op>(call->op_) : nullptr;
-  if (!op || !IsExpectedTpopOp(op->name_, func_type)) return false;
-  if (result_var) *result_var = assign->var_;
-  return true;
-}
-
-bool IsTfreeStmt(const StmtPtr& stmt, VarPtr* tile_var = nullptr, std::string* op_name = nullptr) {
-  auto eval = std::dynamic_pointer_cast<const EvalStmt>(stmt);
-  if (!eval) return false;
-  auto call = std::dynamic_pointer_cast<const Call>(eval->expr_);
-  auto op = call ? std::dynamic_pointer_cast<const Op>(call->op_) : nullptr;
-  if (!op || (op->name_ != "system.tfree_to_aiv" && op->name_ != "system.tfree_to_aic")) {
-    return false;
-  }
-  if (op_name) *op_name = op->name_;
-  if (tile_var) *tile_var = call && !call->args_.empty() ? AsVarLike(call->args_[0]) : nullptr;
-  return true;
-}
-
-std::unordered_set<const Var*> CollectStmtVarRefs(const StmtPtr& stmt) {
-  outline_utils::VarRefCollector refs;
-  refs.VisitStmt(stmt);
-  return refs.var_refs;
-}
-
-std::unordered_set<const Var*> CollectStmtDefinedVars(const StmtPtr& stmt) {
-  std::unordered_set<const Var*> defs;
-  if (auto assign = std::dynamic_pointer_cast<const AssignStmt>(stmt)) {
-    defs.insert(assign->var_.get());
-  } else if (auto for_stmt = std::dynamic_pointer_cast<const ForStmt>(stmt)) {
-    for (const auto& ret : for_stmt->return_vars_) {
-      defs.insert(ret.get());
-    }
-  } else if (auto if_stmt = std::dynamic_pointer_cast<const IfStmt>(stmt)) {
-    for (const auto& ret : if_stmt->return_vars_) {
-      defs.insert(ret.get());
-    }
-  } else if (auto while_stmt = std::dynamic_pointer_cast<const WhileStmt>(stmt)) {
-    for (const auto& ret : while_stmt->return_vars_) {
-      defs.insert(ret.get());
-    }
-  }
-  return defs;
-}
-
-std::unordered_set<const Var*> CollectCallArgVarRefs(const StmtPtr& stmt) {
-  CallPtr call;
-  if (auto assign = std::dynamic_pointer_cast<const AssignStmt>(stmt)) {
-    call = std::dynamic_pointer_cast<const Call>(assign->value_);
-  } else if (auto eval = std::dynamic_pointer_cast<const EvalStmt>(stmt)) {
-    call = std::dynamic_pointer_cast<const Call>(eval->expr_);
-  }
-  if (!call) return CollectStmtVarRefs(stmt);
-
-  std::unordered_set<const Var*> refs_set;
-  for (const auto& arg : call->args_) {
-    outline_utils::VarRefCollector refs;
-    refs.VisitExpr(arg);
-    refs_set.insert(refs.var_refs.begin(), refs.var_refs.end());
-  }
-  return refs_set;
-}
-
-std::vector<const Var*> GetSortedVarRefs(const std::unordered_set<const Var*>& refs) {
-  std::vector<const Var*> sorted_refs(refs.begin(), refs.end());
-  std::sort(sorted_refs.begin(), sorted_refs.end(), [](const Var* lhs, const Var* rhs) {
-    if (lhs == rhs) return false;
-    if (lhs->name_hint_ != rhs->name_hint_) return lhs->name_hint_ < rhs->name_hint_;
-    return lhs->UniqueId() < rhs->UniqueId();
-  });
-  return sorted_refs;
-}
-
-struct TpopChain {
-  size_t tpop_idx;
-  std::vector<size_t> user_idxs;
-  size_t tfree_idx = std::numeric_limits<size_t>::max();
-  VarPtr tpop_var;
-  size_t last_use_idx;
-};
-
-const Var* CanonicalizeTpopRef(const Var* var, const std::unordered_map<const Var*, VarPtr>& tpop_var_remap) {
-  if (!var) return nullptr;
-  auto it = tpop_var_remap.find(var);
-  return (it != tpop_var_remap.end() && it->second) ? it->second.get() : var;
-}
-
-StmtPtr NormalizeNestedTpopChains(const StmtPtr& stmt, CoreSide side,
-                                  const std::unordered_map<const Var*, VarPtr>& tpop_var_remap);
-
-std::vector<StmtPtr> NormalizeTpopChains(const std::vector<StmtPtr>& stmts, CoreSide side,
-                                         const std::unordered_map<const Var*, VarPtr>& tpop_var_remap) {
-  std::vector<StmtPtr> normalized_inputs;
-  normalized_inputs.reserve(stmts.size());
-  for (const auto& stmt : stmts) {
-    normalized_inputs.push_back(NormalizeNestedTpopChains(stmt, side, tpop_var_remap));
-  }
-
-  std::map<const Var*, TpopChain> chains;
-  std::vector<const Var*> tpop_order;
-  std::vector<bool> in_chain(normalized_inputs.size(), false);
-  std::unordered_map<const Var*, size_t> def_indices;
-  bool seen_first_tpop = false;
-
-  for (size_t i = 0; i < normalized_inputs.size(); ++i) {
-    auto assign = std::dynamic_pointer_cast<const AssignStmt>(normalized_inputs[i]);
-    const auto stmt_defs = CollectStmtDefinedVars(normalized_inputs[i]);
-    for (const auto* def : GetSortedVarRefs(stmt_defs)) {
-      def_indices.try_emplace(def, i);
-    }
-    VarPtr tpop_var;
-    if (assign && IsTpopAssignStmt(normalized_inputs[i], &tpop_var)) {
-      seen_first_tpop = true;
-      chains.emplace(tpop_var.get(), TpopChain{i, {}, std::numeric_limits<size_t>::max(), tpop_var, i});
-      tpop_order.push_back(tpop_var.get());
-      in_chain[i] = true;
-      continue;
-    }
-
-    if (!seen_first_tpop) continue;
-
-    VarPtr tfree_var;
-    std::string tfree_op_name;
-    const Var* tfree_key = nullptr;
-    if (IsTfreeStmt(normalized_inputs[i], &tfree_var, &tfree_op_name) && tfree_var &&
-        tfree_op_name == GetTfreeOpName(side)) {
-      tfree_key = CanonicalizeTpopRef(tfree_var.get(), tpop_var_remap);
-    }
-    if (tfree_key && chains.count(tfree_key) > 0) {
-      normalized_inputs[i] = std::make_shared<EvalStmt>(
-          CreateTfree(side, chains.find(tfree_key)->second.tpop_var, normalized_inputs[i]->span_),
-          normalized_inputs[i]->span_);
-      chains.find(tfree_key)->second.tfree_idx = i;
-      in_chain[i] = true;
-      continue;
-    }
-
-    std::unordered_set<const Var*> refs;
-    CallPtr call;
-    if (assign) {
-      call = std::dynamic_pointer_cast<const Call>(assign->value_);
-    } else if (auto eval = std::dynamic_pointer_cast<const EvalStmt>(normalized_inputs[i])) {
-      call = std::dynamic_pointer_cast<const Call>(eval->expr_);
-    }
-    if (call) {
-      auto CollectExprRefs = [&](const ExprPtr& expr) {
-        if (auto var_like = AsVarLike(expr)) {
-          refs.insert(var_like.get());
-          return;
-        }
-        outline_utils::VarRefCollector collector;
-        collector.VisitExpr(expr);
-        refs.insert(collector.var_refs.begin(), collector.var_refs.end());
-      };
-      for (const auto& arg : call->args_) {
-        CollectExprRefs(arg);
-      }
-    } else {
-      refs = CollectStmtVarRefs(normalized_inputs[i]);
-    }
-    const auto sorted_refs = GetSortedVarRefs(refs);
-    for (const auto* ref : sorted_refs) {
-      const Var* canonical_ref = CanonicalizeTpopRef(ref, tpop_var_remap);
-      if (canonical_ref && chains.count(canonical_ref) > 0) {
-        chains.find(canonical_ref)->second.last_use_idx =
-            std::max(chains.find(canonical_ref)->second.last_use_idx, i);
-      }
-    }
-    const Var* referenced_tpop = nullptr;
-    bool has_multi_tpop_refs = false;
-    for (const auto* ref : sorted_refs) {
-      const Var* canonical_ref = CanonicalizeTpopRef(ref, tpop_var_remap);
-      if (canonical_ref && chains.count(canonical_ref) > 0) {
-        if (referenced_tpop && referenced_tpop != canonical_ref) {
-          has_multi_tpop_refs = true;
-          break;
-        }
-        referenced_tpop = canonical_ref;
-      }
-    }
-    bool has_unsafe_dep = false;
-    if (referenced_tpop && !has_multi_tpop_refs) {
-      size_t tpop_idx = chains.find(referenced_tpop)->second.tpop_idx;
-      for (const auto* ref : sorted_refs) {
-        const Var* canonical_ref = CanonicalizeTpopRef(ref, tpop_var_remap);
-        if (canonical_ref == referenced_tpop) continue;
-        auto def_it = def_indices.find(ref);
-        if (def_it != def_indices.end() && def_it->second > tpop_idx && def_it->second < i) {
-          has_unsafe_dep = true;
-          break;
-        }
-      }
-    }
-    if (referenced_tpop && !has_multi_tpop_refs && !has_unsafe_dep) {
-      chains.find(referenced_tpop)->second.user_idxs.push_back(i);
-      in_chain[i] = true;
-    }
-  }
-
-  if (tpop_order.empty()) return normalized_inputs;
-
-  size_t first_tpop = chains[tpop_order[0]].tpop_idx;
-  std::vector<StmtPtr> result;
-  result.reserve(normalized_inputs.size() + tpop_order.size());
-  std::unordered_map<size_t, std::vector<StmtPtr>> deferred_tfrees;
-  for (size_t i = 0; i < first_tpop; ++i) {
-    result.push_back(normalized_inputs[i]);
-  }
-
-  for (const auto* var : tpop_order) {
-    auto& chain = chains[var];
-    result.push_back(normalized_inputs[chain.tpop_idx]);
-    for (size_t user_idx : chain.user_idxs) {
-      result.push_back(normalized_inputs[user_idx]);
-    }
-    size_t last_grouped_idx = chain.user_idxs.empty() ? chain.tpop_idx : chain.user_idxs.back();
-    if (chain.last_use_idx <= last_grouped_idx) {
-      if (chain.tfree_idx != std::numeric_limits<size_t>::max()) {
-        result.push_back(normalized_inputs[chain.tfree_idx]);
-        in_chain[chain.tfree_idx] = true;
-      } else {
-        result.push_back(std::make_shared<EvalStmt>(
-            CreateTfree(side, chain.tpop_var, normalized_inputs[chain.tpop_idx]->span_),
-            normalized_inputs[chain.tpop_idx]->span_));
-      }
-    } else if (chain.tfree_idx == std::numeric_limits<size_t>::max() ||
-               chain.tfree_idx < chain.last_use_idx) {
-      if (chain.tfree_idx != std::numeric_limits<size_t>::max()) {
-        in_chain[chain.tfree_idx] = true;
-      }
-      deferred_tfrees[chain.last_use_idx].push_back(std::make_shared<EvalStmt>(
-          CreateTfree(side, chain.tpop_var, normalized_inputs[chain.tpop_idx]->span_),
-          normalized_inputs[chain.tpop_idx]->span_));
-    } else if (chain.tfree_idx > chain.last_use_idx) {
-      // Keep an existing tfree in its original position when it already follows
-      // the true last use and moving it earlier would be unsafe.
-      continue;
-    } else {
-      deferred_tfrees[chain.last_use_idx].push_back(std::make_shared<EvalStmt>(
-          CreateTfree(side, chain.tpop_var, normalized_inputs[chain.tpop_idx]->span_),
-          normalized_inputs[chain.tpop_idx]->span_));
-      in_chain[chain.tfree_idx] = true;
-    }
-  }
-
-  for (size_t i = first_tpop; i < normalized_inputs.size(); ++i) {
-    if (!in_chain[i]) {
-      result.push_back(normalized_inputs[i]);
-    }
-    auto deferred_it = deferred_tfrees.find(i);
-    if (deferred_it != deferred_tfrees.end()) {
-      result.insert(result.end(), deferred_it->second.begin(), deferred_it->second.end());
-    }
-  }
-  return result;
-}
-
-StmtPtr NormalizeNestedTpopChains(const StmtPtr& stmt, CoreSide side,
-                                  const std::unordered_map<const Var*, VarPtr>& tpop_var_remap) {
-  if (auto for_stmt = std::dynamic_pointer_cast<const ForStmt>(stmt)) {
-    auto new_body = NormalizeTpopChains(FlattenBody(for_stmt->body_), side, tpop_var_remap);
-    return RebuildForStmt(for_stmt, MakeBody(new_body, for_stmt->span_));
-  }
-  if (auto if_stmt = std::dynamic_pointer_cast<const IfStmt>(stmt)) {
-    auto new_then = NormalizeTpopChains(FlattenBody(if_stmt->then_body_), side, tpop_var_remap);
-    std::optional<std::vector<StmtPtr>> new_else;
-    if (if_stmt->else_body_.has_value()) {
-      new_else = NormalizeTpopChains(FlattenBody(if_stmt->else_body_.value()), side, tpop_var_remap);
-    }
-    return RebuildIfStmt(if_stmt, new_then, new_else);
-  }
-  if (auto while_stmt = std::dynamic_pointer_cast<const WhileStmt>(stmt)) {
-    auto new_body = NormalizeTpopChains(FlattenBody(while_stmt->body_), side, tpop_var_remap);
-    return RebuildWhileStmt(while_stmt, MakeBody(new_body, while_stmt->span_));
-  }
-  return stmt;
 }
 
 // ============================================================================
@@ -1915,7 +551,7 @@ ExpandedKernel ExpandMixedFunction(const FunctionPtr& func, bool create_group = 
 
     // Scan original body recursively for tile.store AssignStmts
     std::vector<std::shared_ptr<const AssignStmt>> original_assigns;
-    CollectAllAssignStmts(stmts, original_assigns);
+    dce::CollectAllAssignStmts(stmts, original_assigns);
     for (const auto& assign : original_assigns) {
       auto call = std::dynamic_pointer_cast<const Call>(assign->value_);
       if (!call || !call->op_) continue;
@@ -2252,311 +888,6 @@ Pass ExpandMixedKernel() {
 }
 
 }  // namespace pass
-
-// ============================================================================
-// MixedKernelExpanded property verifier
-// ============================================================================
-
-namespace {
-
-class MixedKernelExpandedVerifier : public IRVisitor {
- public:
-  explicit MixedKernelExpandedVerifier(std::vector<Diagnostic>& diagnostics, std::string func_name)
-      : diagnostics_(diagnostics), func_name_(std::move(func_name)) {}
-
-  void VisitExpr_(const CallPtr& op) override {
-    if (!op || !op->op_) {
-      IRVisitor::VisitExpr_(op);
-      return;
-    }
-    auto affinity = ClassifyCallAffinity(op);
-    if (affinity == CoreAffinity::CUBE) {
-      has_cube_ = true;
-    } else if (affinity == CoreAffinity::VECTOR) {
-      has_vector_ = true;
-    } else if (affinity == CoreAffinity::BOUNDARY) {
-      has_cube_ = true;
-      has_vector_ = true;
-    }
-    IRVisitor::VisitExpr_(op);
-  }
-
-  void CheckResult() {
-    if (has_cube_ && has_vector_) {
-      diagnostics_.emplace_back(DiagnosticSeverity::Error, "MixedKernelExpanded", 0,
-                                "InCore function '" + func_name_ +
-                                    "' contains both Cube and Vector tile ops (should have been expanded)",
-                                Span::unknown());
-    }
-  }
-
- private:
-  std::vector<Diagnostic>& diagnostics_;
-  std::string func_name_;
-  bool has_cube_ = false;
-  bool has_vector_ = false;
-};
-
-class TpopMemoryVerifier : public IRVisitor {
- public:
-  TpopMemoryVerifier(std::vector<Diagnostic>& diagnostics, std::string func_name, FunctionType func_type)
-      : diagnostics_(diagnostics), func_name_(std::move(func_name)), func_type_(func_type) {}
-
-  void VisitStmt_(const AssignStmtPtr& op) override {
-    if (!op) return;
-    auto call = std::dynamic_pointer_cast<const Call>(op->value_);
-    auto ir_op = call ? std::dynamic_pointer_cast<const Op>(call->op_) : nullptr;
-    if (!ir_op) {
-      IRVisitor::VisitStmt_(op);
-      return;
-    }
-
-    std::optional<MemorySpace> expected_memory;
-    if (func_type_ == FunctionType::AIC && ir_op->name_ == "tile.tpop_from_aiv") {
-      expected_memory = MemorySpace::Mat;
-    } else if (func_type_ == FunctionType::AIV && ir_op->name_ == "tile.tpop_from_aic") {
-      expected_memory = MemorySpace::Vec;
-    }
-
-    if (expected_memory.has_value()) {
-      auto tile_type = std::dynamic_pointer_cast<const TileType>(op->var_->GetType());
-      bool valid = tile_type && tile_type->memory_space_.has_value() &&
-                   tile_type->memory_space_.value() == expected_memory.value();
-      if (!valid) {
-        std::string func_kind = (func_type_ == FunctionType::AIC) ? "AIC" : "AIV";
-        std::string actual_memory = (tile_type && tile_type->memory_space_.has_value())
-                                        ? MemorySpaceToString(tile_type->memory_space_.value())
-                                        : "unset";
-        diagnostics_.emplace_back(
-            DiagnosticSeverity::Error, "MixedKernelExpanded", 0,
-            func_kind + " function '" + func_name_ + "' requires " + ir_op->name_ +
-                " result in MemorySpace::" + MemorySpaceToString(expected_memory.value()) +
-                ", got MemorySpace::" + actual_memory,
-            op->span_);
-      }
-    }
-
-    IRVisitor::VisitStmt_(op);
-  }
-
- private:
-  std::vector<Diagnostic>& diagnostics_;
-  std::string func_name_;
-  FunctionType func_type_;
-};
-
-std::string FormatObservedSlotSizes(const std::vector<int64_t>& slot_sizes) {
-  std::string result;
-  for (size_t i = 0; i < slot_sizes.size(); ++i) {
-    if (i > 0) result += ", ";
-    result += std::to_string(slot_sizes[i]);
-  }
-  return result;
-}
-
-void VerifyCrossCorePipeSetup(const FunctionPtr& func, std::vector<Diagnostic>& diagnostics) {
-  CrossCorePipeMetadata metadata;
-  CollectCrossCorePipeMetadata(FlattenBody(func->body_), metadata);
-  if (!metadata.HasCrossCoreOps()) return;
-  CrossCorePipeMetadata dominating_setup = CollectDominatingPipeSetupMetadata(FlattenBody(func->body_));
-
-  auto report_slot_issue = [&](const std::string& issue, const PipeDirectionMetadata& direction,
-                               const std::string& direction_name) {
-    diagnostics.emplace_back(DiagnosticSeverity::Error, "MixedKernelExpanded", 0,
-                             "Function '" + func->name_ + "' uses " + direction_name + " cross-core tiles " +
-                                 issue + ": " + FormatObservedSlotSizes(direction.observed_slot_sizes),
-                             func->span_);
-  };
-
-  if (metadata.c2v.has_inconsistent_slot_size) {
-    report_slot_issue("with inconsistent slot sizes", metadata.c2v, "C2V");
-  }
-  if (metadata.v2c.has_inconsistent_slot_size) {
-    report_slot_issue("with inconsistent slot sizes", metadata.v2c, "V2C");
-  }
-  if ((metadata.c2v.has_ops && !metadata.c2v.slot_size_bytes.has_value()) ||
-      (metadata.v2c.has_ops && !metadata.v2c.slot_size_bytes.has_value())) {
-    diagnostics.emplace_back(
-        DiagnosticSeverity::Error, "MixedKernelExpanded", 0,
-        "Function '" + func->name_ +
-            "' uses cross-core tile ops with non-static tile size; auto pipe setup requires "
-            "statically known tile shapes",
-        func->span_);
-  }
-  if (metadata.c2v.has_ops && metadata.v2c.has_ops && metadata.c2v.slot_size_bytes.has_value() &&
-      metadata.v2c.slot_size_bytes.has_value() &&
-      metadata.c2v.slot_size_bytes.value() != metadata.v2c.slot_size_bytes.value()) {
-    diagnostics.emplace_back(DiagnosticSeverity::Error, "MixedKernelExpanded", 0,
-                             "Function '" + func->name_ +
-                                 "' uses bidirectional cross-core tiles with different "
-                                 "slot sizes (C2V=" +
-                                 std::to_string(metadata.c2v.slot_size_bytes.value()) +
-                                 ", V2C=" + std::to_string(metadata.v2c.slot_size_bytes.value()) +
-                                 "); single initialize_pipe slot_size is unsupported",
-                             func->span_);
-  }
-
-  if (func->func_type_ == FunctionType::AIC) {
-    if (!dominating_setup.has_aic_initialize_pipe) {
-      diagnostics.emplace_back(DiagnosticSeverity::Error, "MixedKernelExpanded", 0,
-                               "AIC function '" + func->name_ +
-                                   "' uses cross-core tile ops but has no 'system.aic_initialize_pipe' call",
-                               func->span_);
-    }
-    if (metadata.v2c.has_ops && !dominating_setup.has_reserve_buffer) {
-      diagnostics.emplace_back(
-          DiagnosticSeverity::Error, "MixedKernelExpanded", 0,
-          "AIC function '" + func->name_ + "' uses V2C cross-core ops but has no 'system.reserve_buffer'",
-          func->span_);
-    }
-    if (metadata.c2v.has_ops && !dominating_setup.has_import_peer_buffer) {
-      diagnostics.emplace_back(
-          DiagnosticSeverity::Error, "MixedKernelExpanded", 0,
-          "AIC function '" + func->name_ + "' uses C2V cross-core ops but has no 'system.import_peer_buffer'",
-          func->span_);
-    }
-  } else if (func->func_type_ == FunctionType::AIV) {
-    if (!dominating_setup.has_aiv_initialize_pipe) {
-      diagnostics.emplace_back(DiagnosticSeverity::Error, "MixedKernelExpanded", 0,
-                               "AIV function '" + func->name_ +
-                                   "' uses cross-core tile ops but has no 'system.aiv_initialize_pipe' call",
-                               func->span_);
-    }
-    if (metadata.c2v.has_ops && !dominating_setup.has_reserve_buffer) {
-      diagnostics.emplace_back(
-          DiagnosticSeverity::Error, "MixedKernelExpanded", 0,
-          "AIV function '" + func->name_ + "' uses C2V cross-core ops but has no 'system.reserve_buffer'",
-          func->span_);
-    }
-    if (metadata.v2c.has_ops && !dominating_setup.has_import_peer_buffer) {
-      diagnostics.emplace_back(
-          DiagnosticSeverity::Error, "MixedKernelExpanded", 0,
-          "AIV function '" + func->name_ + "' uses V2C cross-core ops but has no 'system.import_peer_buffer'",
-          func->span_);
-    }
-  }
-}
-
-bool StmtReferencesVar(const StmtPtr& stmt, const Var* var) {
-  if (!stmt || !var) return false;
-  auto refs = CollectStmtVarRefs(stmt);
-  return refs.count(var) > 0;
-}
-
-void VerifyTpopTfreeOrderInBlock(const std::vector<StmtPtr>& stmts, const FunctionPtr& func,
-                                 std::vector<Diagnostic>& diagnostics);
-
-void VerifyNestedTpopTfreeOrder(const StmtPtr& stmt, const FunctionPtr& func,
-                                std::vector<Diagnostic>& diagnostics) {
-  if (auto for_stmt = std::dynamic_pointer_cast<const ForStmt>(stmt)) {
-    VerifyTpopTfreeOrderInBlock(FlattenBody(for_stmt->body_), func, diagnostics);
-  } else if (auto if_stmt = std::dynamic_pointer_cast<const IfStmt>(stmt)) {
-    VerifyTpopTfreeOrderInBlock(FlattenBody(if_stmt->then_body_), func, diagnostics);
-    if (if_stmt->else_body_.has_value()) {
-      VerifyTpopTfreeOrderInBlock(FlattenBody(if_stmt->else_body_.value()), func, diagnostics);
-    }
-  } else if (auto while_stmt = std::dynamic_pointer_cast<const WhileStmt>(stmt)) {
-    VerifyTpopTfreeOrderInBlock(FlattenBody(while_stmt->body_), func, diagnostics);
-  }
-}
-
-void VerifyTpopTfreeOrderInBlock(const std::vector<StmtPtr>& stmts, const FunctionPtr& func,
-                                 std::vector<Diagnostic>& diagnostics) {
-  const std::string expected_tfree =
-      (func->func_type_ == FunctionType::AIC) ? "system.tfree_to_aiv" : "system.tfree_to_aic";
-  VarPtr open_tpop_var;
-  std::string open_tpop_op_name;
-  const Span* open_tpop_span = &func->span_;
-
-  for (const auto& stmt : stmts) {
-    VerifyNestedTpopTfreeOrder(stmt, func, diagnostics);
-
-    VarPtr tpop_var;
-    if (IsExpectedTpopAssignStmt(stmt, func->func_type_, &tpop_var)) {
-      if (open_tpop_var) {
-        diagnostics.emplace_back(
-            DiagnosticSeverity::Error, "MixedKernelExpanded", 0,
-            "Function '" + func->name_ +
-                "' must order cross-core tpop chains as 'tpop -> use -> tfree -> next tpop'",
-            stmt->span_);
-      }
-      open_tpop_var = tpop_var;
-      open_tpop_span = &stmt->span_;
-      open_tpop_op_name = GetStmtOpName(stmt);
-      continue;
-    }
-
-    VarPtr tfree_var;
-    std::string tfree_op_name;
-    if (IsTfreeStmt(stmt, &tfree_var, &tfree_op_name)) {
-      if (!open_tpop_var) {
-        continue;
-      }
-      if (tfree_op_name != expected_tfree || !tfree_var || tfree_var.get() != open_tpop_var.get()) {
-        diagnostics.emplace_back(DiagnosticSeverity::Error, "MixedKernelExpanded", 0,
-                                 ((func->func_type_ == FunctionType::AIC) ? "AIC" : "AIV") +
-                                     std::string(" function '") + func->name_ + "' must match " +
-                                     open_tpop_op_name + " with '" + expected_tfree +
-                                     "' on the same tile value",
-                                 stmt->span_);
-      } else {
-        open_tpop_var.reset();
-      }
-      continue;
-    }
-
-    if (open_tpop_var && !StmtReferencesVar(stmt, open_tpop_var.get())) {
-      diagnostics.emplace_back(
-          DiagnosticSeverity::Error, "MixedKernelExpanded", 0,
-          "Function '" + func->name_ +
-              "' must order cross-core tpop chains as 'tpop -> use -> tfree -> next tpop'",
-          stmt->span_);
-      open_tpop_var.reset();
-    }
-  }
-
-  if (open_tpop_var) {
-    diagnostics.emplace_back(DiagnosticSeverity::Error, "MixedKernelExpanded", 0,
-                             ((func->func_type_ == FunctionType::AIC) ? "AIC" : "AIV") +
-                                 std::string(" function '") + func->name_ + "' uses " + open_tpop_op_name +
-                                 " but has no matching '" + expected_tfree + "' call",
-                             *open_tpop_span);
-  }
-}
-
-void VerifyTpopTfreeOrder(const FunctionPtr& func, std::vector<Diagnostic>& diagnostics) {
-  VerifyTpopTfreeOrderInBlock(FlattenBody(func->body_), func, diagnostics);
-}
-
-}  // namespace
-
-class MixedKernelExpandedPropertyVerifierImpl : public PropertyVerifier {
- public:
-  [[nodiscard]] std::string GetName() const override { return "MixedKernelExpanded"; }
-
-  void Verify(const ProgramPtr& program, std::vector<Diagnostic>& diagnostics) override {
-    if (!program) return;
-    for (const auto& [gv, func] : program->functions_) {
-      if (!func || !func->body_) continue;
-      if (func->func_type_ == FunctionType::InCore) {
-        MixedKernelExpandedVerifier verifier(diagnostics, func->name_);
-        verifier.VisitStmt(func->body_);
-        verifier.CheckResult();
-        continue;
-      }
-      if (func->func_type_ == FunctionType::AIC || func->func_type_ == FunctionType::AIV) {
-        TpopMemoryVerifier verifier(diagnostics, func->name_, func->func_type_);
-        verifier.VisitStmt(func->body_);
-        VerifyCrossCorePipeSetup(func, diagnostics);
-        VerifyTpopTfreeOrder(func, diagnostics);
-      }
-    }
-  }
-};
-
-PropertyVerifierPtr CreateMixedKernelExpandedPropertyVerifier() {
-  return std::make_shared<MixedKernelExpandedPropertyVerifierImpl>();
-}
 
 }  // namespace ir
 }  // namespace pypto
